# feat: improve adjust apply